Move types for user
