+++ types/users/user.ts
@@ -1,0 +1,26 @@
+export interface IUser {
+    id: number;
+    email: string;
+    username: string;
+    password: string;
+    name: {
+        firstname: string;
+        lastname: string;
+    };
+    address: {
+        city: string;
+        street: string;
+        number: number;
+        zipcode: string;
+        geolocation: {
+            lat: string;
+            long: string;
+        };
+    };
+    phone: string;
+}
+
+export enum SORT {
+    DESC = 'desc',
+    ASC = 'asc'
+}
